+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 0.0.8:
 - `zeus test` now supports running a glob of tests (`zeus test --env testnet my/path/*.sol`)
 - etherscan verification now integrated.
+- Added the "script" phase, which can run a command as part of an upgrade.